libblockfs: Move {File,Node}Operations impl. to separate files
Also move Ext2 code to it's own subdirectory.
