+++ drivers/libblockfs/src/ext2/ext2fs.hpp
@@ -14,7 +14,7 @@
 #include <hel.h>
 
 #include <blockfs.hpp>
-#include "common.hpp"
+#include "../common.hpp"
 #include "fs.bragi.hpp"
 
 namespace blockfs {
