+++ drivers/libblockfs/src/ext2/ops.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <protocols/fs/server.hpp>
+
+namespace blockfs::ext2 {
+
+extern protocols::fs::FileOperations fileOperations;
+extern protocols::fs::NodeOperations nodeOperations;
+
+} // namespace blockfs::ext2
