+++ drivers/libblockfs/src/libblockfs.cpp
@@ -1,7 +1,6 @@
 #include <stdio.h>
 #include <string.h>
 #include <iostream>
-#include <algorithm>
 #include <string>
 #include <sys/epoll.h>
 #include <linux/cdrom.h>
@@ -16,647 +15,16 @@
 
 #include <blockfs.hpp>
 #include "gpt.hpp"
-#include "ext2fs.hpp"
+#include "ext2/ext2fs.hpp"
+#include "ext2/ops.hpp"
 #include "raw.hpp"
+#include "trace.hpp"
 #include "fs.bragi.hpp"
 #include <bragi/helpers-std.hpp>
 
 namespace blockfs {
 
-bool tracingInitialized = false;
 bool clkInitialized = false;
-
-constinit protocols::ostrace::Event ostEvtGetLink{"libblockfs.getLink"};
-constinit protocols::ostrace::Event ostEvtTraverseLinks{"libblockfs.traverseLinks"};
-constinit protocols::ostrace::Event ostEvtRead{"libblockfs.read"};
-constinit protocols::ostrace::Event ostEvtReadDir{"libblockfs.readDir"};
-constinit protocols::ostrace::Event ostEvtWrite{"libblockfs.write"};
-constinit protocols::ostrace::Event ostEvtRawRead{"libblockfs.rawRead"};
-constinit protocols::ostrace::Event ostEvtExt2ManageInode{"ext2.manageInode"};
-constinit protocols::ostrace::Event ostEvtExt2ManageFile{"ext2.manageFile"};
-constinit protocols::ostrace::UintAttribute ostAttrTime{"time"};
-constinit protocols::ostrace::UintAttribute ostAttrNumBytes{"numBytes"};
-
-protocols::ostrace::Vocabulary ostVocabulary{
-	ostEvtGetLink,
-	ostEvtTraverseLinks,
-	ostEvtRead,
-	ostEvtReadDir,
-	ostEvtWrite,
-	ostEvtRawRead,
-	ostEvtExt2ManageInode,
-	ostEvtExt2ManageFile,
-	ostAttrTime,
-	ostAttrNumBytes,
-};
-
-protocols::ostrace::Context ostContext{ostVocabulary};
-
-namespace {
-
-async::result<protocols::fs::SeekResult> seekAbs(void *object, int64_t offset) {
-	auto self = static_cast<ext2fs::OpenFile *>(object);
-	self->offset = offset;
-	co_return static_cast<ssize_t>(self->offset);
-}
-
-async::result<protocols::fs::SeekResult> seekRel(void *object, int64_t offset) {
-	auto self = static_cast<ext2fs::OpenFile *>(object);
-	self->offset += offset;
-	co_return static_cast<ssize_t>(self->offset);
-}
-
-async::result<protocols::fs::SeekResult> seekEof(void *object, int64_t offset) {
-	auto self = static_cast<ext2fs::OpenFile *>(object);
-	self->offset += offset + self->inode->fileSize();
-	co_return static_cast<ssize_t>(self->offset);
-}
-
-using FlockManager = protocols::fs::FlockManager;
-using Flock = protocols::fs::Flock;
-
-async::result<protocols::fs::Error> flock(void *object, int flags) {
-	auto self = static_cast<ext2fs::OpenFile *>(object);
-	co_await self->inode->readyJump.wait();
-	auto inode = self->inode;
-
-	auto result = co_await inode->flockManager.lock(&self->flock, flags);
-	co_return result;
-}
-
-async::result<protocols::fs::ReadResult> read(void *object, helix_ng::CredentialsView,
-		void *buffer, size_t length) {
-	if (!length)
-		co_return size_t{0};
-
-	protocols::ostrace::Timer timer;
-	auto self = static_cast<ext2fs::OpenFile *>(object);
-
-	if(self->inode->fileType == FileType::kTypeDirectory) {
-		co_return protocols::fs::Error::isDirectory;
-	}
-
-	co_await self->inode->readyJump.wait();
-
-	if(self->offset >= self->inode->fileSize())
-		co_return size_t{0};
-
-	auto remaining = self->inode->fileSize() - self->offset;
-	auto chunkSize = std::min(length, remaining);
-	if(!chunkSize)
-		co_return size_t{0}; // TODO: Return an explicit end-of-file error?
-
-	auto chunk_offset = self->offset;
-	self->offset += chunkSize;
-
-	// TODO: If we *know* that the pages are already available,
-	//       we can also fall back to the following "old" mapping code.
-/*
-	auto mapOffset = chunk_offset & ~size_t(0xFFF);
-	auto mapSize = (((chunk_offset & size_t(0xFFF)) + chunkSize + 0xFFF) & ~size_t(0xFFF));
-
-	helix::LockMemoryView lockMemory;
-	auto &&submit = helix::submitLockMemoryView(helix::BorrowedDescriptor(self->inode->frontalMemory),
-			&lockMemory, mapOffset, mapSize, helix::Dispatcher::global());
-	co_await submit.wait();
-	HEL_CHECK(lockMemory.error());
-
-	// Map the page cache into the address space.
-	helix::Mapping fileMap{helix::BorrowedDescriptor{self->inode->frontalMemory},
-			static_cast<ptrdiff_t>(mapOffset), mapSize,
-			kHelMapProtRead | kHelMapDontRequireBacking};
-
-	memcpy(buffer, reinterpret_cast<char *>(fileMap.get()) + (chunk_offset - mapOffset),
-			chunkSize);
-*/
-
-	auto readMemory = co_await helix_ng::readMemory(
-			helix::BorrowedDescriptor(self->inode->frontalMemory),
-			chunk_offset, chunkSize, buffer);
-	HEL_CHECK(readMemory.error());
-
-	ostContext.emit(
-		ostEvtRead,
-		ostAttrNumBytes(length),
-		ostAttrTime(timer.elapsed())
-	);
-
-	co_return chunkSize;
-}
-
-async::result<protocols::fs::ReadResult> pread(void *object, int64_t offset, helix_ng::CredentialsView,
-		void *buffer, size_t length) {
-	assert(length);
-
-	protocols::ostrace::Timer timer;
-	auto self = static_cast<ext2fs::OpenFile *>(object);
-	co_await self->inode->readyJump.wait();
-
-	if(self->offset >= self->inode->fileSize())
-		co_return size_t{0};
-
-	auto remaining = self->inode->fileSize() - offset;
-	auto chunk_size = std::min(length, remaining);
-	if(!chunk_size)
-		co_return size_t{0}; // TODO: Return an explicit end-of-file error?
-
-	auto chunk_offset = offset;
-	auto map_offset = chunk_offset & ~size_t(0xFFF);
-	auto map_size = (((chunk_offset & size_t(0xFFF)) + chunk_size + 0xFFF) & ~size_t(0xFFF));
-
-	helix::LockMemoryView lock_memory;
-	auto &&submit = helix::submitLockMemoryView(helix::BorrowedDescriptor(self->inode->frontalMemory),
-			&lock_memory, map_offset, map_size, helix::Dispatcher::global());
-	co_await submit.async_wait();
-	HEL_CHECK(lock_memory.error());
-
-	// Map the page cache into the address space.
-	helix::Mapping file_map{helix::BorrowedDescriptor{self->inode->frontalMemory},
-			static_cast<ptrdiff_t>(map_offset), map_size,
-			kHelMapProtRead | kHelMapDontRequireBacking};
-
-	memcpy(buffer, reinterpret_cast<char *>(file_map.get()) + (chunk_offset - map_offset),
-			chunk_size);
-
-	ostContext.emit(
-		ostEvtRead,
-		ostAttrNumBytes(length),
-		ostAttrTime(timer.elapsed())
-	);
-
-	co_return chunk_size;
-}
-
-async::result<frg::expected<protocols::fs::Error, size_t>> write(void *object, helix_ng::CredentialsView,
-		const void *buffer, size_t length) {
-	if(!length) {
-		co_return 0;
-	}
-	protocols::ostrace::Timer timer;
-
-	auto self = static_cast<ext2fs::OpenFile *>(object);
-	if(self->append) {
-		self->offset = self->inode->fileSize();
-	}
-	co_await self->inode->fs.write(self->inode.get(), self->offset, buffer, length);
-	self->offset += length;
-
-	ostContext.emit(
-		ostEvtWrite,
-		ostAttrNumBytes(length),
-		ostAttrTime(timer.elapsed())
-	);
-
-	co_return length;
-}
-
-async::result<frg::expected<protocols::fs::Error, size_t>> pwrite(void *object, int64_t offset, helix_ng::CredentialsView credentials,
-			const void *buffer, size_t length) {
-	(void) credentials;
-
-	if(!length) {
-		co_return 0;
-	}
-
-	auto self = static_cast<ext2fs::OpenFile *>(object);
-	co_await self->inode->fs.write(self->inode.get(), offset, buffer, length);
-	co_return length;
-}
-
-async::result<helix::BorrowedDescriptor>
-accessMemory(void *object) {
-	auto self = static_cast<ext2fs::OpenFile *>(object);
-	co_await self->inode->readyJump.wait();
-	co_return self->inode->frontalMemory;
-}
-
-async::result<protocols::fs::ReadEntriesResult>
-readEntries(void *object) {
-	auto self = static_cast<ext2fs::OpenFile *>(object);
-
-	ostContext.emit(
-		ostEvtReadDir
-	);
-
-	co_return co_await self->readEntries();
-}
-
-async::result<frg::expected<protocols::fs::Error>>
-truncate(void *object, size_t size) {
-	auto self = static_cast<ext2fs::OpenFile *>(object);
-	co_await self->inode->fs.truncate(self->inode.get(), size);
-	co_return {};
-}
-
-async::result<int> getFileFlags(void *) {
-	std::cout << "libblockfs: getFileFlags is stubbed" << std::endl;
-    co_return 0;
-}
-
-async::result<void> setFileFlags(void *, int) {
-	std::cout << "libblockfs: setFileFlags is stubbed" << std::endl;
-    co_return;
-}
-
-constexpr protocols::fs::FileOperations fileOperations {
-	.seekAbs      = &seekAbs,
-	.seekRel      = &seekRel,
-	.seekEof      = &seekEof,
-	.read         = &read,
-	.pread        = &pread,
-	.write        = &write,
-	.pwrite       = &pwrite,
-	.readEntries  = &readEntries,
-	.accessMemory = &accessMemory,
-	.truncate     = &truncate,
-	.flock        = &flock,
-	.getFileFlags = &getFileFlags,
-	.setFileFlags = &setFileFlags,
-};
-
-async::result<frg::expected<protocols::fs::Error, protocols::fs::GetLinkResult>>
-getLink(std::shared_ptr<void> object,
-		std::string name) {
-	auto self = std::static_pointer_cast<ext2fs::Inode>(object);
-
-	protocols::ostrace::Timer timer;
-	frg::scope_exit evtOnExit{[&] {
-		ostContext.emit(
-			ostEvtGetLink,
-			ostAttrTime(timer.elapsed())
-		);
-	}};
-
-
-	assert(!name.empty() && name != "." && name != "..");
-	auto entry = FRG_CO_TRY(co_await self->findEntry(name));
-	if(!entry)
-		co_return protocols::fs::GetLinkResult{nullptr, -1,
-				protocols::fs::FileType::unknown};
-
-	protocols::fs::FileType type;
-	switch(entry->fileType) {
-	case kTypeDirectory:
-		type = protocols::fs::FileType::directory;
-		break;
-	case kTypeRegular:
-		type = protocols::fs::FileType::regular;
-		break;
-	case kTypeSymlink:
-		type = protocols::fs::FileType::symlink;
-		break;
-	default:
-		throw std::runtime_error("Unexpected file type");
-	}
-
-	assert(entry->inode);
-	co_return protocols::fs::GetLinkResult{self->fs.accessInode(entry->inode), entry->inode, type};
-}
-
-async::result<protocols::fs::GetLinkResult> link(std::shared_ptr<void> object,
-		std::string name, int64_t ino) {
-	auto self = std::static_pointer_cast<ext2fs::Inode>(object);
-	auto entry = co_await self->link(std::move(name), ino, kTypeRegular);
-	if(!entry)
-		co_return protocols::fs::GetLinkResult{nullptr, -1,
-				protocols::fs::FileType::unknown};
-
-	protocols::fs::FileType type;
-	switch(entry->fileType) {
-	case kTypeDirectory:
-		type = protocols::fs::FileType::directory;
-		break;
-	case kTypeRegular:
-		type = protocols::fs::FileType::regular;
-		break;
-	case kTypeSymlink:
-		type = protocols::fs::FileType::symlink;
-		break;
-	default:
-		throw std::runtime_error("Unexpected file type");
-	}
-
-	assert(entry->inode);
-	co_return protocols::fs::GetLinkResult{self->fs.accessInode(entry->inode), entry->inode, type};
-}
-
-async::result<frg::expected<protocols::fs::Error>> unlink(std::shared_ptr<void> object, std::string name) {
-	auto self = std::static_pointer_cast<ext2fs::Inode>(object);
-	auto result = co_await self->unlink(std::move(name));
-	if(!result) {
-		assert(result.error() == protocols::fs::Error::fileNotFound
-			|| result.error() == protocols::fs::Error::directoryNotEmpty);
-		co_return result.error();
-	}
-	co_return {};
-}
-
-async::detached serve(smarter::shared_ptr<ext2fs::OpenFile> file,
-		helix::UniqueLane local_ctrl, helix::UniqueLane local_pt) {
-	async::cancellation_event cancel_pt;
-
-	// Cancel the passthrough lane once the file line is closed.
-	async::detach(protocols::fs::serveFile(std::move(local_ctrl),
-			file.get(), &fileOperations), [&] {
-		cancel_pt.cancel();
-	});
-
-	co_await protocols::fs::servePassthrough(std::move(local_pt),
-			file, &fileOperations, cancel_pt);
-}
-
-async::result<protocols::fs::FileStats>
-getStats(std::shared_ptr<void> object) {
-	auto self = std::static_pointer_cast<ext2fs::Inode>(object);
-	co_await self->readyJump.wait();
-
-	protocols::fs::FileStats stats;
-	stats.linkCount = self->diskInode()->linksCount;
-	stats.fileSize = self->fileSize();
-	stats.mode = self->diskInode()->mode & 0xFFF;
-	stats.uid = self->uid;
-	stats.gid = self->gid;
-	stats.accessTime.tv_sec = self->diskInode()->atime;
-	stats.dataModifyTime.tv_sec = self->diskInode()->mtime;;
-	stats.anyChangeTime.tv_sec = self->diskInode()->ctime;
-
-	co_return stats;
-}
-
-async::result<protocols::fs::OpenResult>
-open(std::shared_ptr<void> object, bool append) {
-	auto self = std::static_pointer_cast<ext2fs::Inode>(object);
-	auto file = smarter::make_shared<ext2fs::OpenFile>(self);
-	co_await self->readyJump.wait();
-	file->append = append;
-
-	helix::UniqueLane local_ctrl, remote_ctrl;
-	helix::UniqueLane local_pt, remote_pt;
-	std::tie(local_ctrl, remote_ctrl) = helix::createStream();
-	std::tie(local_pt, remote_pt) = helix::createStream();
-	struct timespec time = clk::getRealtime();
-	self->diskInode()->atime = time.tv_sec;
-
-	auto syncInode = co_await helix_ng::synchronizeSpace(
-			helix::BorrowedDescriptor{kHelNullHandle},
-			self->diskMapping.get(), self->fs.inodeSize);
-	HEL_CHECK(syncInode.error());
-
-	serve(file, std::move(local_ctrl), std::move(local_pt));
-
-	co_return protocols::fs::OpenResult{std::move(remote_ctrl), std::move(remote_pt)};
-}
-
-async::result<std::string> readSymlink(std::shared_ptr<void> object) {
-	auto self = std::static_pointer_cast<ext2fs::Inode>(object);
-	co_await self->readyJump.wait();
-
-	if(self->fileSize() <= 60) {
-		co_return std::string{self->diskInode()->data.embedded,
-			self->diskInode()->data.embedded + self->fileSize()};
-	} else {
-		std::string result;
-		result.resize(self->fileSize());
-		co_await helix_ng::readMemory(
-			helix::BorrowedDescriptor(self->frontalMemory),
-			0, self->fileSize(), result.data());
-		co_return result;
-	}
-}
-
-async::result<protocols::fs::MkdirResult>
-mkdir(std::shared_ptr<void> object, std::string name) {
-	auto self = std::static_pointer_cast<ext2fs::Inode>(object);
-	auto entry = co_await self->mkdir(std::move(name));
-
-	if(!entry)
-		co_return protocols::fs::MkdirResult{nullptr, -1};
-
-	assert(entry->inode);
-	co_return protocols::fs::MkdirResult{self->fs.accessInode(entry->inode), entry->inode};
-}
-
-async::result<protocols::fs::SymlinkResult>
-symlink(std::shared_ptr<void> object, std::string name, std::string target) {
-	auto self = std::static_pointer_cast<ext2fs::Inode>(object);
-	auto entry = co_await self->symlink(std::move(name), std::move(target));
-
-	if(!entry)
-		co_return protocols::fs::SymlinkResult{nullptr, -1};
-
-	assert(entry->inode);
-	co_return protocols::fs::SymlinkResult{self->fs.accessInode(entry->inode), entry->inode};
-}
-
-async::result<protocols::fs::Error> chmod(std::shared_ptr<void> object, int mode) {
-	auto self = std::static_pointer_cast<ext2fs::Inode>(object);
-	auto result = co_await self->chmod(mode);
-
-	co_return result;
-}
-
-async::result<protocols::fs::Error> utimensat(std::shared_ptr<void> object,
-		std::optional<timespec> atime, std::optional<timespec> mtime, timespec ctime) {
-	auto self = std::static_pointer_cast<ext2fs::Inode>(object);
-	auto result = co_await self->utimensat(atime, mtime, ctime);
-
-	co_return result;
-}
-
-async::result<void> obstructLink(std::shared_ptr<void> object, std::string name) {
-	auto self = std::static_pointer_cast<ext2fs::Inode>(object);
-	self->obstructedLinks.insert(name);
-	co_return;
-}
-
-async::result<protocols::fs::TraverseLinksResult> traverseLinks(std::shared_ptr<void> object,
-		std::deque<std::string> components) {
-	auto self = std::static_pointer_cast<ext2fs::Inode>(object);
-
-	protocols::ostrace::Timer timer;
-	frg::scope_exit evtOnExit{[&] {
-		ostContext.emit(
-			ostEvtTraverseLinks,
-			ostAttrTime(timer.elapsed())
-		);
-	}};
-
-	std::optional<ext2fs::DirEntry> entry;
-	std::shared_ptr<ext2fs::Inode> parent = self;
-	size_t processedComponents = 0;
-
-	std::vector<std::pair<std::shared_ptr<void>, int64_t>> nodes;
-
-	while (!components.empty()) {
-		auto component = components.front();
-		components.pop_front();
-		processedComponents++;
-
-		if (component == "..") {
-			if (parent == self)
-				co_return std::make_tuple(nodes, protocols::fs::FileType::unknown, 0);
-
-			parent = self->fs.accessInode(FRG_CO_TRY(co_await parent->findEntry(".."))->inode);
-			nodes.pop_back();
-		} else {
-			entry = FRG_CO_TRY(co_await parent->findEntry(component));
-
-			if (!entry) {
-				co_return protocols::fs::Error::fileNotFound;
-			}
-
-			assert(entry->inode);
-			nodes.push_back({self->fs.accessInode(entry->inode), entry->inode});
-
-			if (!components.empty()) {
-				if (parent->obstructedLinks.find(component) != parent->obstructedLinks.end()) {
-					break;
-				}
-
-				auto ino = self->fs.accessInode(entry->inode);
-				if (entry->fileType == kTypeSymlink)
-					break;
-
-				if (entry->fileType != kTypeDirectory)
-					co_return protocols::fs::Error::notDirectory;
-
-				parent = ino;
-			}
-		}
-	}
-
-	if(!entry)
-		co_return protocols::fs::Error::fileNotFound;
-
-	protocols::fs::FileType type;
-	switch(entry->fileType) {
-	case kTypeDirectory:
-		type = protocols::fs::FileType::directory;
-		break;
-	case kTypeRegular:
-		type = protocols::fs::FileType::regular;
-		break;
-	case kTypeSymlink:
-		type = protocols::fs::FileType::symlink;
-		break;
-	default:
-		throw std::runtime_error("Unexpected file type");
-	}
-
-	co_return std::make_tuple(nodes, type, processedComponents);
-}
-
-constexpr protocols::fs::NodeOperations nodeOperations{
-	.getStats = &getStats,
-	.getLink = &getLink,
-	.link = &link,
-	.unlink = &unlink,
-	.open = &open,
-	.readSymlink = &readSymlink,
-	.mkdir = &mkdir,
-	.symlink = &symlink,
-	.chmod = &chmod,
-	.utimensat = &utimensat,
-	.obstructLink = &obstructLink,
-	.traverseLinks = &traverseLinks
-};
-
-async::result<protocols::fs::ReadResult> rawRead(void *object, helix_ng::CredentialsView,
-		void *buffer, size_t length) {
-	assert(length);
-
-	uint64_t start;
-	HEL_CHECK(helGetClock(&start));
-
-	auto self = static_cast<raw::OpenFile *>(object);
-	auto file_size = co_await self->rawFs->device->getSize();
-
-	if(self->offset >= file_size)
-		co_return size_t{0};
-
-	auto remaining = file_size - self->offset;
-	auto chunkSize = std::min(length, remaining);
-	if(!chunkSize)
-		co_return size_t{0}; // TODO: Return an explicit end-of-file error?
-
-	auto chunk_offset = self->offset;
-	self->offset += chunkSize;
-
-	auto readMemory = co_await helix_ng::readMemory(
-			helix::BorrowedDescriptor(self->rawFs->frontalMemory),
-			chunk_offset, chunkSize, buffer);
-	HEL_CHECK(readMemory.error());
-
-	uint64_t end;
-	HEL_CHECK(helGetClock(&end));
-
-	ostContext.emit(
-		ostEvtRawRead,
-		ostAttrNumBytes(length),
-		ostAttrTime(end - start)
-	);
-
-	co_return chunkSize;
-}
-
-async::result<protocols::fs::Error> rawFlock(void *object, int flags) {
-	auto self = static_cast<raw::OpenFile*>(object);
-
-	auto result = co_await self->rawFs->flockManager.lock(&self->flock, flags);
-	co_return result;
-}
-
-async::result<protocols::fs::SeekResult> rawSeekAbs(void *object, int64_t offset) {
-	auto self = static_cast<raw::OpenFile*>(object);
-	self->offset = offset;
-	co_return static_cast<ssize_t>(self->offset);
-}
-
-async::result<protocols::fs::SeekResult> rawSeekRel(void *object, int64_t offset) {
-	auto self = static_cast<raw::OpenFile*>(object);
-	self->offset += offset;
-	co_return static_cast<ssize_t>(self->offset);
-}
-
-async::result<protocols::fs::SeekResult> rawSeekEof(void *object, int64_t offset) {
-	auto self = static_cast<raw::OpenFile *>(object);
-	auto size = co_await self->rawFs->device->getSize();
-	self->offset = offset + size;
-	co_return static_cast<ssize_t>(self->offset);
-}
-async::result<void> rawIoctl(void *object, uint32_t id, helix_ng::RecvInlineResult msg,
-		helix::UniqueLane conversation) {
-	auto self = static_cast<raw::OpenFile *>(object);
-
-	if(id == managarm::fs::GenericIoctlRequest::message_id) {
-		auto req = bragi::parse_head_only<managarm::fs::GenericIoctlRequest>(msg);
-		assert(req);
-		if (req->command() == CDROM_GET_CAPABILITY) {
-			managarm::fs::GenericIoctlReply rsp;
-			rsp.set_error(managarm::fs::Errors::NOT_A_TERMINAL);
-
-			auto ser = rsp.SerializeAsString();
-			auto [send_resp] = co_await helix_ng::exchangeMsgs(
-					conversation,
-					helix_ng::sendBuffer(ser.data(), ser.size())
-			);
-			HEL_CHECK(send_resp.error());
-		} else {
-			co_await self->rawFs->device->handleIoctl(req.value(), std::move(conversation));
-		}
-	}
-}
-
-constexpr protocols::fs::FileOperations rawOperations {
-	.seekAbs = rawSeekAbs,
-	.seekRel = rawSeekRel,
-	.seekEof = rawSeekEof,
-	.read = rawRead,
-	.ioctl = rawIoctl,
-	.flock = rawFlock,
-};
-
-} // anonymous namespace
 
 BlockDevice::BlockDevice(size_t sector_size, int64_t parent_id)
 : size(0), sectorSize(sector_size), parentId(parent_id) { }
@@ -711,7 +79,7 @@ async::detached servePartition(helix::UniqueLane lane, gpt::Partition *partition
 			helix::UniqueLane local_lane, remote_lane;
 			std::tie(local_lane, remote_lane) = helix::createStream();
 			protocols::fs::serveNode(std::move(local_lane), fs->accessRoot(),
-					&nodeOperations);
+					&ext2::nodeOperations);
 
 			managarm::fs::SvrResponse resp;
 			resp.set_error(managarm::fs::Errors::SUCCESS);
@@ -730,7 +98,7 @@ async::detached servePartition(helix::UniqueLane lane, gpt::Partition *partition
 			helix::UniqueLane local_lane, remote_lane;
 			std::tie(local_lane, remote_lane) = helix::createStream();
 			protocols::fs::serveNode(std::move(local_lane),
-					inode, &nodeOperations);
+					inode, &ext2::nodeOperations);
 
 			managarm::fs::SvrResponse resp;
 			resp.set_error(managarm::fs::Errors::SUCCESS);
@@ -839,8 +207,8 @@ async::detached servePartition(helix::UniqueLane lane, gpt::Partition *partition
 			std::tie(local_lane, remote_lane) = helix::createStream();
 			auto file = smarter::make_shared<raw::OpenFile>(rawFs.get());
 			async::detach(protocols::fs::servePassthrough(std::move(local_lane),
-					file,
-					&rawOperations));
+							file,
+							&raw::rawOperations));
 
 			managarm::fs::SvrResponse resp;
 			resp.set_error(managarm::fs::Errors::SUCCESS);
@@ -944,8 +312,8 @@ async::detached serveDevice(helix::UniqueLane lane, std::unique_ptr<raw::RawFs> 
 			std::tie(local_lane, remote_lane) = helix::createStream();
 			auto file = smarter::make_shared<raw::OpenFile>(rawFs.get());
 			async::detach(protocols::fs::servePassthrough(std::move(local_lane),
-					file,
-					&rawOperations));
+							file,
+							&raw::rawOperations));
 
 			managarm::fs::SvrResponse resp;
 			resp.set_error(managarm::fs::Errors::SUCCESS);
